Got it ^_^
I deleted the setFont for the old Panel Attack thing on welcome screen, deleted the welcome to thing, changed the frame size just a little, moved the background label for the game behind the buttons to they would show up on top of the background if things got weird...and added the gif. Not much, but I can confidently push and make changes now without having to go to my local copy.